Fixed content offset issue after reload
Sorry for the late change. But i was playing around with the demo app and figured when you scrolled down in the content list sometimes the content offset is not being updated in case the new content size is smaller than the one before the reload. So I changed the code to always scroll to top after a reload.